Jc datamodelfix (#59)
* attempt to fix delete FAQ

* tiny fix

* attempt again to fix delete FAQ

+++ src/main/java/com/example/cs4500_sp19_josies/services/FAQService.java
@@ -51,10 +51,12 @@ public class FAQService {
 
   @DeleteMapping("/api/faqs/{faqId}")
   public void deleteFrequentlyAskedQuestion(
-          @PathVariable("faqId") Integer id,
-          @RequestBody FrequentlyAskedQuestion faqUpdates) {
-    for (FrequentlyAskedAnswer faa : faqUpdates.getAnswers()) {
-      answerRepo.deleteById(faa.getId());
+          @PathVariable("faqId") Integer id) {
+    FrequentlyAskedQuestion faq = repository.findFrequentlyAskedQuestionById(id);
+    if (faq.getAnswers() != null) {
+      for (FrequentlyAskedAnswer faa : faq.getAnswers()) {
+        answerRepo.deleteById(faa.getId());
+      }
     }
     repository.deleteById(id);
   }
